[sw/device/lib/dif] Prepare to add C warning flags

+++ sw/device/lib/dif/dif_aes.c
@@ -173,7 +173,7 @@ static dif_result_t configure_aux(const dif_aes_t *aes,
 static void aes_set_multireg(const dif_aes_t *aes, const uint32_t *data,
                              size_t regs_num, ptrdiff_t reg0_offset) {
   for (int i = 0; i < regs_num; ++i) {
-    ptrdiff_t offset = reg0_offset + (i * sizeof(uint32_t));
+    ptrdiff_t offset = reg0_offset + (ptrdiff_t)i * (ptrdiff_t)sizeof(uint32_t);
 
     mmio_region_write32(aes->base_addr, offset, data[i]);
   }
@@ -182,7 +182,7 @@ static void aes_set_multireg(const dif_aes_t *aes, const uint32_t *data,
 static void aes_read_multireg(const dif_aes_t *aes, uint32_t *data,
                               size_t regs_num, ptrdiff_t reg0_offset) {
   for (int i = 0; i < regs_num; ++i) {
-    ptrdiff_t offset = reg0_offset + (i * sizeof(uint32_t));
+    ptrdiff_t offset = reg0_offset + (ptrdiff_t)i * (ptrdiff_t)sizeof(uint32_t);
 
     data[i] = mmio_region_read32(aes->base_addr, offset);
   }
@@ -410,7 +410,8 @@ dif_result_t dif_aes_read_iv(const dif_aes_t *aes, dif_aes_iv_t *iv) {
   }
 
   for (int i = 0; i < AES_IV_MULTIREG_COUNT; ++i) {
-    ptrdiff_t offset = AES_IV_0_REG_OFFSET + (i * sizeof(uint32_t));
+    ptrdiff_t offset =
+        AES_IV_0_REG_OFFSET + (ptrdiff_t)i * (ptrdiff_t)sizeof(uint32_t);
 
     iv->iv[i] = mmio_region_read32(aes->base_addr, offset);
   }
