+++ sw/device/lib/dif/dif_entropy_src.c
@@ -346,13 +346,13 @@ dif_result_t dif_entropy_src_get_health_test_stats(
         return kDifError;
     }
 
-    stats->high_watermark[i] =
-        mmio_region_read32(entropy_src->base_addr, high_watermarks_reg_offset);
+    stats->high_watermark[i] = (uint16_t)mmio_region_read32(
+        entropy_src->base_addr, high_watermarks_reg_offset);
     stats->low_watermark[i] =
         low_watermarks_reg_offset == -1
             ? 0
-            : mmio_region_read32(entropy_src->base_addr,
-                                 low_watermarks_reg_offset);
+            : (uint16_t)mmio_region_read32(entropy_src->base_addr,
+                                           low_watermarks_reg_offset);
 
     stats->high_fails[i] =
         mmio_region_read32(entropy_src->base_addr, high_fails_reg_offset);
@@ -372,7 +372,7 @@ dif_result_t dif_entropy_src_get_alert_fail_counts(
     return kDifBadArg;
   }
 
-  counts->total_fails = mmio_region_read32(
+  counts->total_fails = (uint16_t)mmio_region_read32(
       entropy_src->base_addr, ENTROPY_SRC_ALERT_SUMMARY_FAIL_COUNTS_REG_OFFSET);
 
   uint32_t alert_fail_counts = mmio_region_read32(
@@ -381,37 +381,40 @@ dif_result_t dif_entropy_src_get_alert_fail_counts(
       entropy_src->base_addr, ENTROPY_SRC_EXTHT_FAIL_COUNTS_REG_OFFSET);
 
   // Unpack high threshold failure counts.
-  counts->high_fails[kDifEntropySrcTestRepetitionCount] = bitfield_field32_read(
-      alert_fail_counts, ENTROPY_SRC_ALERT_FAIL_COUNTS_REPCNT_FAIL_COUNT_FIELD);
+  counts->high_fails[kDifEntropySrcTestRepetitionCount] =
+      (uint8_t)bitfield_field32_read(
+          alert_fail_counts,
+          ENTROPY_SRC_ALERT_FAIL_COUNTS_REPCNT_FAIL_COUNT_FIELD);
   counts->high_fails[kDifEntropySrcTestRepetitionCountSymbol] =
-      bitfield_field32_read(
+      (uint8_t)bitfield_field32_read(
           alert_fail_counts,
           ENTROPY_SRC_ALERT_FAIL_COUNTS_REPCNTS_FAIL_COUNT_FIELD);
   counts->high_fails[kDifEntropySrcTestAdaptiveProportion] =
-      bitfield_field32_read(
+      (uint8_t)bitfield_field32_read(
           alert_fail_counts,
           ENTROPY_SRC_ALERT_FAIL_COUNTS_ADAPTP_HI_FAIL_COUNT_FIELD);
-  counts->high_fails[kDifEntropySrcTestBucket] = bitfield_field32_read(
+  counts->high_fails[kDifEntropySrcTestBucket] = (uint8_t)bitfield_field32_read(
       alert_fail_counts, ENTROPY_SRC_ALERT_FAIL_COUNTS_BUCKET_FAIL_COUNT_FIELD);
-  counts->high_fails[kDifEntropySrcTestMarkov] = bitfield_field32_read(
+  counts->high_fails[kDifEntropySrcTestMarkov] = (uint8_t)bitfield_field32_read(
       alert_fail_counts,
       ENTROPY_SRC_ALERT_FAIL_COUNTS_MARKOV_HI_FAIL_COUNT_FIELD);
-  counts->high_fails[kDifEntropySrcTestMailbox] = bitfield_field32_read(
-      extht_alert_fail_counts,
-      ENTROPY_SRC_EXTHT_FAIL_COUNTS_EXTHT_HI_FAIL_COUNT_FIELD);
+  counts->high_fails[kDifEntropySrcTestMailbox] =
+      (uint8_t)bitfield_field32_read(
+          extht_alert_fail_counts,
+          ENTROPY_SRC_EXTHT_FAIL_COUNTS_EXTHT_HI_FAIL_COUNT_FIELD);
 
   // Unpack low threshold failure counts.
   counts->low_fails[kDifEntropySrcTestRepetitionCount] = 0;
   counts->low_fails[kDifEntropySrcTestRepetitionCountSymbol] = 0;
   counts->low_fails[kDifEntropySrcTestAdaptiveProportion] =
-      bitfield_field32_read(
+      (uint8_t)bitfield_field32_read(
           alert_fail_counts,
           ENTROPY_SRC_ALERT_FAIL_COUNTS_ADAPTP_LO_FAIL_COUNT_FIELD);
   counts->low_fails[kDifEntropySrcTestBucket] = 0;
-  counts->low_fails[kDifEntropySrcTestMarkov] = bitfield_field32_read(
+  counts->low_fails[kDifEntropySrcTestMarkov] = (uint8_t)bitfield_field32_read(
       alert_fail_counts,
       ENTROPY_SRC_ALERT_FAIL_COUNTS_MARKOV_LO_FAIL_COUNT_FIELD);
-  counts->low_fails[kDifEntropySrcTestMailbox] = bitfield_field32_read(
+  counts->low_fails[kDifEntropySrcTestMailbox] = (uint8_t)bitfield_field32_read(
       extht_alert_fail_counts,
       ENTROPY_SRC_EXTHT_FAIL_COUNTS_EXTHT_LO_FAIL_COUNT_FIELD);
 
@@ -639,7 +642,7 @@ dif_result_t dif_entropy_src_get_debug_state(
 
   uint32_t debug_state_reg = mmio_region_read32(
       entropy_src->base_addr, ENTROPY_SRC_DEBUG_STATUS_REG_OFFSET);
-  debug_state->entropy_fifo_depth = bitfield_field32_read(
+  debug_state->entropy_fifo_depth = (uint8_t)bitfield_field32_read(
       debug_state_reg, ENTROPY_SRC_DEBUG_STATUS_ENTROPY_FIFO_DEPTH_FIELD);
   debug_state->sha3_fsm_state = bitfield_field32_read(
       debug_state_reg, ENTROPY_SRC_DEBUG_STATUS_SHA3_FSM_FIELD);
